requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-Caching==2.1.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.12.1
redis==5.0.1

# Asynchronous Tasks
celery==5.3.6
flower==2.0.1

# Web Scraping
selenium==4.16.0
playwright==1.41.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0
Pillow==10.1.0
webdriver-manager==4.0.1
fake-useragent==1.4.0

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-magic==0.4.27

# AI and Data Processing
anthropic==0.12.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
gensim==4.3.2

# Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
validators==0.22.0
tenacity==8.2.3
pydantic==2.5.3
structlog==23.2.0
jsonschema==4.20.0

# Production
gunicorn==21.2.0
eventlet==0.35.0
gevent==23.9.1